Added tests for the MathJax Arabic extension and its CDN
