+++ common/djangoapps/edraak_tests/__init__.py
@@ -1,0 +1,3 @@
+"""
+A module for adding system-wide tests for Edraak.
+"""
